graph update, naming conventions to match optimizer suite
