+++ README.md
@@ -99,10 +99,10 @@ In the particle_swarm.py class, the objective function is called twice. Some opt
 ## Example Implementations
 
 ### Basic PSO Example
-pso_test.py provides a sample use case of the optimizer. 
+main_test.py provides a sample use case of the optimizer. 
 
 ### Detailed Messages
-pso_test_details.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. This implementation sets up the hooks for integration with AntennaCAT in order to provide the user feedback of warnings and errors.
+main_test_details.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. This implementation sets up the hooks for integration with AntennaCAT in order to provide the user feedback of warnings and errors.
 
 ### Realtime Graph
 
@@ -110,7 +110,7 @@ pso_test_details.py provides an example using a parent class, and the self.suppr
         <img src="https://github.com/LC-Linkous/pso_python/blob/main/media/pso_graph.gif" alt="Example PSO Convergence" height="200">
 </p>
 
-pso_test_graph.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. Additionally, a realtime graph shows particle locations at every step.
+main_test_graph.py provides an example using a parent class, and the self.suppress_output and detailedWarnings flags to control error messages that are passed back to the parent class to be printed with a timestamp. Additionally, a realtime graph shows particle locations at every step.
 
 NOTE: if you close the graph as the code is running, the code will continue to run, but the graph will not re-open.
